Fixed PM problem

+++ createSoftLinks.sh
@@ -1,18 +1,32 @@
 #!/bin/bash
 dirs='examples/Decode examples/Encode examples/Incoming'
+if [ $# -gt 0 ]
+then
+	crdirs=true
+else
+	crdirs=false
+fi
+#echo $crdirs
 current=$PWD
 for d in $dirs
 do
-	#echo $d
-	#echo $current
 	cd $current
 	cd $d
 	cd lib
-	rm -r pdulib
-	mkdir pdulib
-	cd pdulib
-	ln -s ../../../../src/pdulib.cpp pdulib.cpp
-	ln -s ../../../../src/pdulib.h pdulib.h
+	if [ -d pdulib ]
+	then
+		rm -r pdulib
+	fi
+	if [[ $crdirs = true ]]
+	then
+		echo "Creating symlinks"
+		mkdir pdulib
+		cd pdulib
+		ln -s ../../../../src/pdulib.cpp pdulib.cpp
+		ln -s ../../../../src/pdulib.h pdulib.h
+	else
+		echo "Not creating symlinks"
+	fi
 done
 echo 'All done'
